requirements.txt
certifi==2022.9.24
charset-normalizer==2.1.1
idna==3.4
isort==5.10.1
numpy==1.23.4
pandas==1.5.1
psycopg2-binary==2.9.4
python-dateutil==2.8.2
python-dotenv==0.21.0
pytz==2022.5
requests==2.28.1
six==1.16.0
urllib3==1.26.12
